Cloud Functions/weather_measurement_transformation/requirements.txt
functions-framework==3.*
pandas==2.1.*
firebase-functions
firebase-admin
google-cloud-storage